requirements.txt
Django==1.10.3
requests==2.11.1
djangorestframework==3.3.3
mongoengine==0.9
pymongo==2.8
django-rest-framework-mongoengine
